use python directly
https://coveord.atlassian.net/browse/CDX-885
