+++ packages/cli-e2e/mitmproxy/main.py
@@ -1,0 +1,3 @@
+from mitmproxy.tools.main import mitmdump
+
+mitmdump()
