+++ packages/cli-e2e/mitmproxy/requirements.txt
@@ -1,0 +1,1 @@
+mitmproxy
