+++ packages/cli-e2e/package.json
@@ -17,6 +17,7 @@
     "url": "git+https://github.com/coveo/cli.git"
   },
   "scripts": {
+    "postinstall": "pip install -r mitmproxy/requirements.txt",
     "lint": "eslint .",
     "test-e2e": "npm run test:headless",
     "test:headless": "node ./entrypoints/entry.js",
